refactor: split dce switch pruning
